@@ -4,47 +4,49 @@
   "private": true,
   "dependencies": {
     "@capacitor/app": "4.0.1",
-    "@capacitor/camera": "^4.1.1",
-    "@capacitor/core": "4.2.0",
+    "@capacitor/camera": "^4.1.2",
+    "@capacitor/core": "4.3.0",
     "@capacitor/haptics": "4.0.1",
     "@capacitor/keyboard": "4.0.1",
     "@capacitor/network": "^4.0.1",
     "@capacitor/preferences": "^4.0.1",
-    "@capacitor/splash-screen": "^4.0.1",
+    "@capacitor/splash-screen": "^4.1.0",
     "@capacitor/status-bar": "4.0.1",
-    "@ionic/react": "^6.0.0",
-    "@ionic/react-router": "^6.0.0",
+    "@ionic/react": "^6.2.9",
+    "@ionic/react-router": "^6.2.9",
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^13.3.0",
-    "@testing-library/user-event": "^12.6.3",
-    "@types/jest": "^26.0.20",
-    "@types/node": "^12.19.15",
-    "@types/react": "^18.0.17",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.1.1",
+    "@types/node": "^18.8.2",
+    "@types/react": "^18.0.21",
     "@types/react-dom": "^18.0.6",
     "@types/react-router": "^5.1.11",
     "@types/react-router-dom": "^5.1.7",
-    "bootstrap": "5.2.1",
+    "bootstrap": "5.2.2",
     "ionicons": "^6.0.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
+    "react-form-hook": "^0.0.1",
+    "react-query": "^3.39.2",
+    "react-router": "^6.4.1",
+    "react-router-dom": "^6.4.1",
     "react-scripts": "^5.0.0",
     "recoil": "^0.7.5",
-    "typescript": "^4.1.3",
-    "web-vitals": "^0.2.4",
-    "workbox-background-sync": "^5.1.4",
-    "workbox-broadcast-update": "^5.1.4",
-    "workbox-cacheable-response": "^5.1.4",
-    "workbox-core": "^5.1.4",
-    "workbox-expiration": "^5.1.4",
-    "workbox-google-analytics": "^5.1.4",
-    "workbox-navigation-preload": "^5.1.4",
-    "workbox-precaching": "^5.1.4",
-    "workbox-range-requests": "^5.1.4",
-    "workbox-routing": "^5.1.4",
-    "workbox-strategies": "^5.1.4",
-    "workbox-streams": "^5.1.4"
+    "typescript": "^4.8.4",
+    "web-vitals": "^3.0.3",
+    "workbox-background-sync": "^6.5.4",
+    "workbox-broadcast-update": "^6.5.4",
+    "workbox-cacheable-response": "^6.5.4",
+    "workbox-core": "^6.5.4",
+    "workbox-expiration": "^6.5.4",
+    "workbox-google-analytics": "^6.5.4",
+    "workbox-navigation-preload": "^6.5.4",
+    "workbox-precaching": "^6.5.4",
+    "workbox-range-requests": "^6.5.4",
+    "workbox-routing": "^6.5.4",
+    "workbox-strategies": "^6.5.4",
+    "workbox-streams": "^6.5.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -71,7 +73,7 @@
     ]
   },
   "devDependencies": {
-    "@capacitor/cli": "4.2.0"
+    "@capacitor/cli": "4.3.0"
   },
   "description": "An Ionic project"
 }
